Allow to use immutable maps as an input for Copy.expand
The underlying SimpleTemplateEngine modifies the map that is used as
a parameter. This might be problematic if the argument is an immutable,
e. g. Collections.emptyMap() or some Kotlin collection. The added
`out` variable may also confuse users when it appears in the error
message.

This CL prepares a mutable copy to pass to the SimpleTemplateEngine so
original map is not modified.